Integrate evaluation

+++ computers/windows/docker/scripts/install.bat
@@ -15,31 +15,41 @@ echo ================================================= >> "%LOGFILE%"
 echo Installation started at %date% %time% >> "%LOGFILE%"
 echo ================================================= >> "%LOGFILE%"
 
+echo Setting up winget links for PATH >> "%LOGFILE%"
+setx PATH "%PATH%;%LOCALAPPDATA%\Microsoft\WinGet\Links" /M >nul
+
 REM ---------------------------
 REM 2) Download & Install Python system-wide (silent)
 REM ---------------------------
 set "PYTHON_INSTALLER=%TEMP%\python_installer.exe"
 set "PYTHON_URL=https://www.python.org/ftp/python/3.10.0/python-3.10.0-amd64.exe"
 
-echo Downloading Python installer... >> "%LOGFILE%"
-curl -L -o "%PYTHON_INSTALLER%" "%PYTHON_URL%" >> "%LOGFILE%" 2>&1
-
-echo Installing Python system-wide... >> "%LOGFILE%"
-"%PYTHON_INSTALLER%" /quiet InstallAllUsers=1 PrependPath=1 InstallLauncherAllUsers=1 >> "%LOGFILE%" 2>&1
-if %ERRORLEVEL% neq 0 (
-    echo Python installation failed with error code %ERRORLEVEL%. >> "%LOGFILE%"
-    exit /b %ERRORLEVEL%
-)
-
+REM Define Python paths
 set "PYTHON_PATH=C:\Program Files\Python310\python.exe"
 set "PYTHONW_PATH=C:\Program Files\Python310\pythonw.exe"
+
+REM Check if Python is already installed
+if exist %PYTHON_PATH% (
+    echo Python is already installed at %PYTHON_PATH%. Skipping installation. >> "%LOGFILE%"
+) else (
+    echo Python not found. Proceeding with installation. >> "%LOGFILE%"
+    
+    echo Downloading Python installer... >> "%LOGFILE%"
+    curl -L -o "%PYTHON_INSTALLER%" "%PYTHON_URL%" >> "%LOGFILE%" 2>&1
+
+    echo Installing Python system-wide... >> "%LOGFILE%"
+    "%PYTHON_INSTALLER%" /quiet InstallAllUsers=1 PrependPath=1 InstallLauncherAllUsers=1 >> "%LOGFILE%" 2>&1
+    if %ERRORLEVEL% neq 0 (
+        echo Python installation failed with error code %ERRORLEVEL%. >> "%LOGFILE%"
+        exit /b %ERRORLEVEL%
+    )
+)
+
 setx PYTHON "%PYTHON_PATH%" /M >nul
 setx PYTHONW "%PYTHONW_PATH%" /M >nul
 
 echo Python set to %PYTHON_PATH% >> "%LOGFILE%"
 echo Pythonw set to %PYTHONW_PATH% >> "%LOGFILE%"
-
-
 
 REM ---------------------------
 REM 3) Install software
@@ -71,12 +81,23 @@ echo Installing Python libraries for 'server browser control' ... >> "%LOGFILE%"
 "%PYTHON_PATH%" -m pip install -r "\\host.lan\Data\server_browser_control\requirements.txt" >> "%LOGFILE%" 2>&1
 echo Python libraries for 'server browser control' were installed >> "%LOGFILE%"
 
+echo Installing Python libraries for 'server network proxy' ... >> "%LOGFILE%"
+"%PYTHON_PATH%" -m pip install -r "\\host.lan\Data\server_network_proxy\requirements.txt" >> "%LOGFILE%" 2>&1
+echo Python libraries for 'server network proxy' were installed >> "%LOGFILE%"
+
+echo Installing Python libraries for 'server evaluator' ... >> "%LOGFILE%"
+"%PYTHON_PATH%" -m pip install -r "\\host.lan\Data\server_evaluator\requirements.txt" >> "%LOGFILE%" 2>&1
+echo Python libraries for 'server evaluator' were installed >> "%LOGFILE%"
+
+
 REM ---------------------------
 REM 5) Add Firewall Rules
 REM ---------------------------
 echo Adding firewall rules... >> "%LOGFILE%"
 netsh advfirewall firewall add rule name="SERVER_COMPUTER_CONTROL" dir=in action=allow protocol=TCP localport=5050
 netsh advfirewall firewall add rule name="SERVER_BROWSER_CONTROL" dir=in action=allow protocol=TCP localport=5051
+netsh advfirewall firewall add rule name="SERVER_NETWORK_PROXY" dir=in action=allow protocol=TCP localport=5052
+netsh advfirewall firewall add rule name="SERVER_EVALUATOR" dir=in action=allow protocol=TCP localport=5053
 REM Add firewall rule for Chrome DevTools (port 9222)
 netsh advfirewall firewall add rule name="Chrome Remote Debugging Port" dir=in action=allow protocol=TCP localport=9222
 echo Firewall rules added >> "%LOGFILE%"
@@ -96,8 +117,19 @@ set "STARTUP_SERVER_BROWSER_CONTROL_BAT=%~dp0start_server_browser_control.bat"
     echo @echo off
     echo start /b "" "%PYTHONW_PATH%" "\\host.lan\Data\server_browser_control\server.py"
 ) > "%STARTUP_SERVER_BROWSER_CONTROL_BAT%"
-echo .bat files for servers created >> "%LOGFILE%"
 
+set "STARTUP_SERVER_NETWORK_PROXY_BAT=%~dp0start_server_network_proxy.bat"
+(
+    echo @echo off
+    echo start /b "" "%PYTHONW_PATH%" "\\host.lan\Data\server_network_proxy\server.py"
+) > "%STARTUP_SERVER_NETWORK_PROXY_BAT%"
+
+set "STARTUP_SERVER_EVALUATOR_BAT=%~dp0start_server_evaluator.bat"
+(
+    echo @echo off
+    echo start /b "" "%PYTHONW_PATH%" "\\host.lan\Data\server_evaluator\server.py"
+) > "%STARTUP_SERVER_EVALUATOR_BAT%"
+echo .bat files for servers created >> "%LOGFILE%"
 
 REM ---------------------------
 REM 7) Schedule Startup Task
@@ -107,11 +139,15 @@ REM Without /DELAY is needed in order to wait until network storage is available
 echo Creating 'scheduled tasks' for servers >> "%LOGFILE%"
 schtasks /Create /TN "StartServer-ComputerControl" /SC ONSTART /TR "\"%STARTUP_SERVER_COMPUTER_CONTROL_BAT%\"" /RU "%USERNAME%" /RL HIGHEST /IT /DELAY 0000:30 /F
 schtasks /Create /TN "StartServer-BrowserControl" /SC ONSTART /TR "\"%STARTUP_SERVER_BROWSER_CONTROL_BAT%\"" /RU "%USERNAME%" /RL HIGHEST /IT /DELAY 0000:30 /F
+schtasks /Create /TN "StartServer-NetworkProxy" /SC ONSTART /TR "\"%STARTUP_SERVER_NETWORK_PROXY_BAT%\"" /RU "%USERNAME%" /RL HIGHEST /IT /DELAY 0000:30 /F
+schtasks /Create /TN "StartServer-Evaluator" /SC ONSTART /TR "\"%STARTUP_SERVER_EVALUATOR_BAT%\"" /RU "%USERNAME%" /RL HIGHEST /IT /DELAY 0000:30 /F
 echo 'Scheduled tasks' for servers created >> "%LOGFILE%"
 
 echo Triggering 'scheduled tasks' for servers >> "%LOGFILE%"
 schtasks /Run /TN "StartServer-ComputerControl"
 schtasks /Run /TN "StartServer-BrowserControl"
+schtasks /Run /TN "StartServer-NetworkProxy"
+schtasks /Run /TN "StartServer-Evaluator"
 echo 'Scheduled tasks' for servers started >> "%LOGFILE%"
 
 echo Installation completed at %date% %time% >> "%LOGFILE%"
